requirements.txt
django
django-ckeditor
django-haystack
whoosh
